# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: CornerIQ API
  version: 1.0.0
  description: Convenience store management platform API
servers:
  - url: http://localhost:3001/api
components:
  schemas:
...
    Supplier:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        vatNumber:
          type: string
          nullable: true
        paymentTermsDays:
          type: integer
          nullable: true
        leadTimeDays:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateSupplierBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        vatNumber:
          type: string
        paymentTermsDays:
          type: integer
        leadTimeDays:
          type: integer
        notes:
          type: string
...
paths:
...
  /suppliers:
    get:
      tags: [suppliers]
      operationId: getSuppliers
      summary: List suppliers
      responses:
        '200':
          description: Suppliers list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Supplier'
    post:
      tags: [suppliers]
      operationId: createSupplier
      summary: Create supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierBody'
      responses:
        '201':
          description: Created supplier
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Supplier'
  /suppliers/{id}:
    patch:
      tags: [suppliers]
      operationId: updateSupplier
      summary: Update supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierBody'
      responses:
        '200':
          description: Updated supplier
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Supplier'
...